Pulled out Routes Configuration
* Created a separate class for configuring routes
* This makes it easy to change out if another user wants to have
different routes
