+++ src/main/java/http_server/Main.java
@@ -8,11 +8,13 @@ public class Main {
 
   public static void main(String[] args) throws Exception {
     Router router = new Router();
-    Configuration configuration = new Configuration(router);
-    configuration.parse(args);
-    configuration.populateRoutes();
+    ConfigurationCommandLine configurationCommandLine = new ConfigurationCommandLine(router);
+    configurationCommandLine.parse(args);
+    String directoryPath = configurationCommandLine.getDirectoryName();
+    ConfigurationRoutes configurationRoutes = new ConfigurationRoutes(router, directoryPath);
+    configurationRoutes.populateRoutes();
 
-    ServerSocket serverSocket = new ServerSocket(configuration.getPortNumber());
+    ServerSocket serverSocket = new ServerSocket(configurationCommandLine.getPortNumber());
     ConnectionManager server = new Server(serverSocket);
 
     ExecutorService threadPool = Executors.newFixedThreadPool(4);
