--- src/test/java/http_server/ConfigurationTest.java
@@ -1,83 +1,0 @@
-package http_server;
-
-import org.junit.Test;
-
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.InputStreamReader;
-
-import static org.junit.Assert.*;
-
-public class ConfigurationTest {
-  private String rootPath = System.getProperty("user.dir");
-
-  private Response generateHelloWorldResponse(Router router) throws Exception {
-    String request = "GET /hello_world HTTP/1.1\r\nHost: localhost\r\n\r\n";
-    ByteArrayInputStream inputStream = new ByteArrayInputStream(request.getBytes());
-    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-    return router.generateHttpResponse(bufferedReader);
-  }
-
-
-  @Test
-  public void directoryAndPortAreParsed(){
-    Router router = new Router();
-    Configuration configuration = new Configuration(router);
-
-    String[] commandLineArgs = {"-p", "5000", "-d", rootPath + "/new_directory"};
-    configuration.parse(commandLineArgs);
-
-    assertEquals(rootPath + "/new_directory", configuration.getDirectoryName());
-    assertEquals(new Integer(5000), configuration.getPortNumber());
-
-  }
-
-  @Test
-  public void defaultDirectoryPathIsReturned() {
-    Router router = new Router();
-    Configuration configuration = new Configuration(router);
-
-    String[] commandLineArgs = {"-p", "5000"};
-    configuration.parse(commandLineArgs);
-
-    assertEquals(rootPath + "/code", configuration.getDirectoryName());
-    assertEquals(new Integer(5000), configuration.getPortNumber());
-  }
-
-  @Test
-  public void defaultPortIsReturned() {
-    Router router = new Router();
-    Configuration configuration = new Configuration(router);
-
-    String[] commandLineArgs = {"-d", rootPath + "/new_directory"};
-    configuration.parse(commandLineArgs);
-
-    assertEquals(rootPath + "/new_directory", configuration.getDirectoryName());
-    assertEquals(new Integer(4444), configuration.getPortNumber());
-  }
-
-  @Test
-  public void defaultDirectoryPathAndPortAreReturned() {
-    Router router = new Router();
-    Configuration configuration = new Configuration(router);
-
-    String[] commandLineArgs = {};
-    configuration.parse(commandLineArgs);
-
-    assertEquals(rootPath + "/code", configuration.getDirectoryName());
-    assertEquals(new Integer(4444), configuration.getPortNumber());
-  }
-
-  @Test
-  public void routesArePopulated() throws Exception {
-    Router router = new Router();
-    Configuration configuration = new Configuration(router);
-
-    String[] commandLineArgs = {"-d", rootPath +"/code"};
-    configuration.parse(commandLineArgs);
-    configuration.populateRoutes();
-
-    Response actualResponse = generateHelloWorldResponse(router);
-    assertEquals("hello world", new String(actualResponse.getBody()));
-  }
-}
